+++ README.md
@@ -2,3 +2,4 @@
 
 - 增加 step 18: 增加「第一张图片」和「最后一张张图片」的模态提示。
 - 增加 step 19: 当在网格模式下点击图片时候，自动切换到 slideshow 模式。
+- 增加 step 20: 将不同的组件分离出单个模块文件。
