+++ src/components/image_grid.rs
@@ -1,0 +1,53 @@
+use makepad_widgets::*;
+use crate::app::State;
+
+
+live_design! {
+    use link::widgets::*;
+    use crate::components::image_row::ImageRow;
+
+    pub ImageGrid = {{ImageGrid}} {
+        <PortalList> {
+            flow: Down,
+            ImageRow = <ImageRow> {}
+        }
+    }
+}
+
+
+#[derive(Live, LiveHook, Widget)]
+pub struct ImageGrid {
+    #[deref]
+    view: View,
+}
+
+impl Widget for ImageGrid {
+    fn draw_walk(
+        &mut self,
+        cx: &mut Cx2d,
+        scope: &mut Scope,
+        walk: Walk,
+    ) -> DrawStep {
+        while let Some(item) = self.view.draw_walk(cx, scope, walk).step() {
+            if let Some(mut list) = item.as_portal_list().borrow_mut() {
+                let state = scope.data.get_mut::<State>().unwrap();
+
+                list.set_item_range(cx, 0, state.num_rows());
+                while let Some(row_idx) = list.next_visible_item(cx) {
+                    if row_idx >= state.num_rows() {
+                        continue;
+                    }
+
+                    let row = list.item(cx, row_idx, live_id!(ImageRow));
+                    let mut scope = Scope::with_data_props(state, &row_idx);
+                    row.draw_all(cx, &mut scope);
+                }
+            }
+        }
+        DrawStep::done()
+    }
+
+    fn handle_event(&mut self, cx: &mut Cx, event: &Event, scope: &mut Scope) {
+        self.view.handle_event(cx, event, scope)
+    }
+}
