+++ src/components/mod.rs
@@ -1,0 +1,14 @@
+use makepad_widgets::*;
+
+
+pub mod image_item;
+pub mod image_row;
+pub mod image_grid;
+
+pub use image_item::ImageClickedAction;
+
+pub fn live_design(cx: &mut Cx) {
+    image_item::live_design(cx);
+    image_row::live_design(cx);
+    image_grid::live_design(cx);
+}
